add a new namespace "sliceable"
