+++ chainer/datasets/sliceable/__init__.py
@@ -1,0 +1,5 @@
+from chainer.datasets.sliceable.sliceable_dataset import SliceableDataset  # NOQA
+from chainer.datasets.sliceable.sliceable_getter_dataset import SliceableGetterDataset  # NOQA
+
+from chainer.datasets.sliceable.concatenate import concatenate  # NOQA
+from chainer.datasets.sliceable.transform import transform  # NOQA
