--- tests/chainer_tests/datasets_tests/test_pickable_dataset.py
@@ -1,110 +1,0 @@
-import unittest
-
-from chainer.datasets import PickableDataset
-from chainer import testing
-
-
-class SampleDataset(PickableDataset):
-    def __init__(self, len):
-        super(SampleDataset, self).__init__()
-
-        self.add_getter(self.get_image, 'img')
-        self.add_getter(self.get_anno0_anno2, ('anno0', 'anno2'))
-        self.add_getter(self.get_anno1, 'anno1')
-        self.keys = ('img', 'anno0', 'anno1', 'anno2')
-
-        self.len = len
-        self.count = 0
-
-    def __len__(self):
-        return self.len
-
-    def get_image(self, i):
-        self.count += 1
-        return 'img({:d})'.format(i)
-
-    def get_anno0_anno2(self, i):
-        self.count += 1
-        return 'anno0({:d})'.format(i), 'anno2({:d})'.format(i)
-
-    def get_anno1(self, i):
-        self.count += 1
-        return 'anno1({:d})'.format(i)
-
-
-class TestPickableDataset(unittest.TestCase):
-
-    def setUp(self):
-        self.len = 10
-        self.dataset = SampleDataset(self.len)
-
-    def test_base(self):
-        self.assertEqual(len(self.dataset), self.len)
-        self.assertEqual(
-            self.dataset[0], ('img(0)', 'anno0(0)', 'anno1(0)', 'anno2(0)'))
-        self.assertEqual(self.dataset.count, 3)
-
-    def test_slice_keys_single(self):
-        dataset = self.dataset.slice[:, 'anno0']
-        self.assertEqual(len(dataset), self.len)
-        self.assertEqual(dataset[1], 'anno0(1)')
-        self.assertEqual(self.dataset.count, 1)
-
-    def test_slice_keys_single_tuple(self):
-        dataset = self.dataset.slice[:, ('anno1',)]
-        self.assertEqual(len(dataset), self.len)
-        self.assertEqual(dataset[2], ('anno1(2)',))
-        self.assertEqual(self.dataset.count, 1)
-
-    def test_slice_keys_multiple(self):
-        dataset = self.dataset.slice[:, ('anno0', 'anno2')]
-        self.assertEqual(len(dataset), self.len)
-        self.assertEqual(dataset[3], ('anno0(3)', 'anno2(3)'))
-        self.assertEqual(self.dataset.count, 1)
-
-    def test_slice_keys_invalid(self):
-        with self.assertRaises(KeyError):
-            self.dataset.slice[:, 'invalid']
-
-    def test_slice_index_slice(self):
-        dataset = self.dataset.slice[3:8:2]
-        self.assertEqual(len(dataset), 3)
-        self.assertEqual(
-            dataset[1], ('img(5)', 'anno0(5)', 'anno1(5)', 'anno2(5)'))
-        self.assertEqual(self.dataset.count, 3)
-
-    def test_slice_index_list(self):
-        dataset = self.dataset.slice[[2, 1, 5]]
-        self.assertEqual(len(dataset), 3)
-        self.assertEqual(
-            dataset[0], ('img(2)', 'anno0(2)', 'anno1(2)', 'anno2(2)'))
-        self.assertEqual(self.dataset.count, 3)
-
-    def test_concat(self):
-        dataset = self.dataset.concat(self.dataset)
-        self.assertEqual(len(dataset), 20)
-        self.assertEqual(
-            dataset[1], ('img(1)', 'anno0(1)', 'anno1(1)', 'anno2(1)'))
-        self.assertEqual(
-            dataset[11], ('img(1)', 'anno0(1)', 'anno1(1)', 'anno2(1)'))
-        self.assertEqual(self.dataset.count, 6)
-
-    def test_concat_invalid(self):
-        dataset0 = self.dataset.slice[:, ('img', 'anno0')]
-        dataset1 = self.dataset.slice[:, ('img', 'anno1')]
-        with self.assertRaises(ValueError):
-            dataset0.concat(dataset1)
-
-    def test_transform(self):
-        def transform(in_data):
-            img, anno0, anno1, anno2 = in_data
-            return 'transformed_' + img, 'transformed_' + anno1
-
-        dataset = self.dataset.transform(transform, ('img', 'anno1'))
-        self.assertEqual(len(dataset), self.len)
-        self.assertEqual(
-            dataset[3], ('transformed_img(3)', 'transformed_anno1(3)'))
-        self.assertEqual(self.dataset.count, 3)
-
-
-testing.run_module(__name__, __file__)
